fakeguard/requirements.txt
torch>=2.0.0
transformers>=4.38.0
datasets>=2.18.0
scikit-learn>=1.4.0
spacy>=3.7.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
pyyaml>=6.0.0
